requirements.txt
django
django-tinymce
djangorestframework
django-cleanup
openpyxl
qrcode
django-social-share
khayyam
pillow
pusher